refactor: menu 리팩터링
