+++ src/main/java/kitchenpos/menus/domain/MenuProduct.java
@@ -1,64 +1,47 @@
 package kitchenpos.menus.domain;
 
 import javax.persistence.*;
+import java.util.Objects;
 import java.util.UUID;
 
-import kitchenpos.products.domain.Product;
-
-@Table(name = "menu_product")
-@Entity
+@Embeddable
 public class MenuProduct {
     @Column(name = "seq")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     private Long seq;
 
-    @ManyToOne(optional = false)
-    @JoinColumn(
-        name = "product_id",
-        columnDefinition = "binary(16)",
-        foreignKey = @ForeignKey(name = "fk_menu_product_to_product")
-    )
-    private Product product;
-
-    @Column(name = "quantity", nullable = false)
-    private long quantity;
-
-    @Transient
     private UUID productId;
 
-    public MenuProduct() {
+    @Column(name = "quantity", nullable = false)
+    @Embedded
+    private Quantity quantity;
+
+    private Price price;
+
+    protected MenuProduct() {
     }
 
-    public Long getSeq() {
-        return seq;
-    }
-
-    public void setSeq(final Long seq) {
-        this.seq = seq;
-    }
-
-    public Product getProduct() {
-        return product;
-    }
-
-    public void setProduct(final Product product) {
-        this.product = product;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(final long quantity) {
-        this.quantity = quantity;
-    }
-
-    public UUID getProductId() {
-        return productId;
-    }
-
-    public void setProductId(final UUID productId) {
+    public MenuProduct(UUID productId, Quantity quantity, Price productPrice) {
         this.productId = productId;
+        this.quantity = quantity;
+        this.price = productPrice.multiplyQuantity(quantity);
+    }
+
+    public Price getPrice() {
+        return price;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        MenuProduct that = (MenuProduct) o;
+        return Objects.equals(productId, that.productId) && Objects.equals(quantity, that.quantity);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(productId, quantity);
     }
 }
