+++ src/main/java/kitchenpos/menus/domain/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.domain;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
