+++ src/main/java/kitchenpos/products/application/ChangePriceRequest.java
@@ -1,0 +1,11 @@
+package kitchenpos.products.application;
+
+import kitchenpos.products.domain.Price;
+
+public class ChangePriceRequest {
+    private Price price;
+
+    public Price getPrice() {
+        return price;
+    }
+}
