+++ src/main/java/kitchenpos/products/ui/ProductRestController.java
@@ -1,5 +1,6 @@
 package kitchenpos.products.ui;
 
+import kitchenpos.products.application.ChangePriceRequest;
 import kitchenpos.products.application.CreateProductRequest;
 import kitchenpos.products.application.ProductService;
 import kitchenpos.products.domain.Product;
@@ -27,7 +28,7 @@ public class ProductRestController {
     }
 
     @PutMapping("/{productId}/price")
-    public ResponseEntity<Product> changePrice(@PathVariable final UUID productId, @RequestBody final Product request) {
+    public ResponseEntity<Product> changePrice(@PathVariable final UUID productId, @RequestBody final ChangePriceRequest request) {
         return ResponseEntity.ok(productService.changePrice(productId, request));
     }
