+++ src/test/java/kitchenpos/menus/tobe/domain/DisplayedNameTest.java
@@ -1,5 +1,7 @@
 package kitchenpos.menus.tobe.domain;
 
+import kitchenpos.menus.domain.DisplayedName;
+import kitchenpos.menus.domain.PurgomalumClient;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
