+++ src/test/java/kitchenpos/menus/tobe/domain/MenuProductTest.java
@@ -1,5 +1,7 @@
 package kitchenpos.menus.tobe.domain;
 
+import kitchenpos.menus.domain.Price;
+import kitchenpos.menus.domain.Quantity;
 import org.junit.jupiter.api.Test;
 
 import java.util.UUID;
