+++ src/test/java/kitchenpos/menus/tobe/domain/MenuProductsTest.java
@@ -1,5 +1,8 @@
 package kitchenpos.menus.tobe.domain;
 
+import kitchenpos.menus.domain.MenuProducts;
+import kitchenpos.menus.domain.Price;
+import kitchenpos.menus.domain.Quantity;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
